rename importer to case_importer_v1
and add case_importer_v2 stub
