+++ corehq/apps/case_importer_v1/management/commands/import_cases.py
@@ -2,8 +2,8 @@ import json
 from datetime import datetime
 from django.core.management import BaseCommand, CommandError
 from dimagi.utils.web import json_handler
-from corehq.apps.importer.tasks import do_import
-from corehq.apps.importer.util import ImporterConfig, ExcelFile
+from corehq.apps.case_importer_v1.tasks import do_import
+from corehq.apps.case_importer_v1.util import ImporterConfig, ExcelFile
 from corehq.apps.users.models import WebUser
