+++ corehq/apps/case_importer_v1/templates/case_importer_v1/partials/import_status.html
