+++ corehq/apps/case_importer_v1/util.py
@@ -6,9 +6,9 @@ import xlrd
 from django.utils.translation import ugettext_lazy as _
 from couchdbkit import NoResultFound
 
-from corehq.apps.importer.const import LookupErrors, ImportErrors
+from corehq.apps.case_importer_v1.const import LookupErrors, ImportErrors
 from corehq.apps.groups.models import Group
-from corehq.apps.importer.exceptions import (
+from corehq.apps.case_importer_v1.exceptions import (
     ImporterExcelFileEncrypted,
     ImporterExcelError,
     ImporterFileNotFound,
