+++ settings.py
@@ -290,7 +290,7 @@ HQ_APPS = (
     'corehq.apps.app_manager',
     'corehq.apps.es',
     'corehq.apps.fixtures',
-    'corehq.apps.importer',
+    'corehq.apps.case_importer_v1',
     'corehq.apps.reminders',
     'corehq.apps.translations',
     'corehq.apps.tour',
@@ -1239,7 +1239,7 @@ COUCHDB_APPS = [
     'hqcase',
     'hqmedia',
     'hope',
-    'importer',
+    'case_importer_v1',
     'indicators',
     'locations',
     'mobile_auth',
