+++ urls.py
@@ -64,7 +64,7 @@ domain_specific = patterns('',
     (r'^case/', include('corehq.apps.case_search.urls')),
     (r'^cloudcare/', include('corehq.apps.cloudcare.urls')),
     (r'^fixtures/', include('corehq.apps.fixtures.urls')),
-    (r'^importer/', include('corehq.apps.importer.urls')),
+    (r'^importer/', include('corehq.apps.case_importer_v1.urls')),
     (r'^fri/', include('custom.fri.urls')),
     (r'^ilsgateway/', include('custom.ilsgateway.urls')),
     (r'^ewsghana/', include('custom.ewsghana.urls')),
